Adjust upload of files to new API
The upload of files completely changed and now it seems that a
WeTransfer Storm API and S3 upload are used.

Adjust the code to call this API (based on reverse engineering of link
upload transfer).

At the moment the eu-west-1 region is hardcoded but we can probably use
other regions as well but some way should be investigated to retrieve
such regions and use the nearest one (a big XXX comment documents
that).

Tested only with a single file upload, both with link and email
uploads.

Should close issue #38.