Retry tests with an exponential backoff mechanism
Unfortunately tests are pretty flacky... Sometimes Tor is not ready
other times probably WeTransfer is not happy about us.

Try to introduce an exponential backoff mechanism and retry tests
several times to possibly workaround that.

This could be a bad idea!: i.e. maybe it will just fail several times
in a row! (but maybe not, let's try to use more brute-force in that
case!)